Config PIDs

+++ lib-ws28xx/include/pixelconfiguration.h
@@ -41,11 +41,11 @@ public:
 		return m_type;
 	}
 
-	void SetCount(uint32_t nCount) {
+	void SetCount(uint16_t nCount) {
 		m_nCount = nCount == 0 ? pixel::defaults::COUNT : nCount;
 	}
 
-	uint32_t GetCount() const {
+	uint16_t GetCount() const {
 		return m_nCount;
 	}
 
@@ -123,7 +123,7 @@ public:
 
 private:
 	pixel::Type m_type { pixel::defaults::TYPE };
-	uint32_t m_nCount { pixel::defaults::COUNT };
+	uint16_t m_nCount { pixel::defaults::COUNT };
 	pixel::Map m_map { pixel::Map::UNDEFINED };
 	uint32_t m_nClockSpeedHz { 0 };
 	uint8_t m_nLowCode { 0 };
